new features

+++ HiggsAnalysis/VBFHiggsToWWto2l2nu/python/VBFEventSelector_cfi.py
@@ -3,6 +3,6 @@ import FWCore.ParameterSet.Config as cms
 selectedEvents = cms.EDFilter("VBFEventSelector",
       tagJetInputTag = cms.InputTag("jetTagger:taggedJets"),
       jetDeltaEtaMin = cms.double(5),
-      jetPtMin = cms.double(40),
-      MjjMin = cms.double(500)
+      jetPtMin = cms.double(15),
+      MjjMin = cms.double(40)
      )
